# api/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Viz API
  version: 0.1.0
  description: OpenAPI specification for the Viz API (chi-based)
servers:
  - url: http://{host}:{port}
    description: Local API
    variables:
      host:
        default: localhost
        description: Hostname or IP for the API server
      port:
        default: "7770"
        description: Port for the API server
  - url: /api
    description: Production API
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /auth/apikey:
    get:
      summary: Generate an API key
      operationId: generateApiKey
      security: []
      responses:
        "200":
          description: API key generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKey"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api-keys:
    post:
      summary: Create a new API key
      operationId: createApiKey
      security:
        - BearerAuth: [apikeys:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/APIKeyCreate"
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreateResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      summary: List API keys for the authenticated user
      operationId: listApiKeys
      security:
        - BearerAuth: [apikeys:read]
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api-keys/{uid}:
    get:
      summary: Get API key details
      operationId: getApiKey
      security:
        - BearerAuth: [apikeys:read]
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: API key details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKey"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete an API key
      operationId: deleteApiKey
      security:
        - BearerAuth: [apikeys:write]
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api-keys/{uid}/revoke:
    post:
      summary: Revoke an API key
      operationId: revokeApiKey
      security:
        - BearerAuth: [apikeys:write]
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api-keys/{uid}/rotate:
    post:
      summary: Rotate an API key (revoke old, create new)
      operationId: rotateApiKey
      security:
        - BearerAuth: [apikeys:write]
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: New API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreateResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "API key for authentication. The key should be provided in the Authorization header as a Bearer token. This token is a hex-encoded string of random bytes, prefixed with 'IMAG_'. Bearer format is intentionally omitted to avoid JWT confusion."
      x-scopes:
        admin:read: Read admin resources
        admin:write: Write admin resources
        api_keys:create: Create API keys
        api_keys:list: List API keys
        api_keys:get: Get API key details
        api_keys:revoke: Revoke API keys
        api_keys:rotate: Rotate API keys
        api_keys:delete: Delete API keys
        auth:refresh: Refresh authentication tokens
        auth:revoke: Revoke authentication tokens
        collections:create: Create collections
        collections:read: Read collections
        collections:update: Update collections
        collections:delete: Delete collections
        collections:share: Share collections
        images:read: Read images
        images:upload: Upload images
        images:update: Update images
        images:delete: Delete images
        images:download: Download images
        jobs:read: Read jobs
        jobs:create: Create jobs
        jobs:update: Update jobs
        jobs:delete: Delete jobs
        users:create: Create users
        users:read: Read users
        users:update: Update users
        users:delete: Delete users
        user_settings:read: Read user settings
        user_settings:update: Update user settings
        downloads:create: Create downloads
        events:read: Read events
        "*": Grants all permissions (superuser)

    CookieAuth:
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          description: Response message
      required: [message]
...
    User:
      x-entity: true
      type: object
      properties:
        uid: { type: string, description: User UID }
        first_name: { type: string, description: First name }
        last_name: { type: string, description: Last name }
        username: { type: string, description: Username }
        email: { type: string, description: Email }
        role:
          type: string
          enum: [user, admin, superadmin, guest]
          description: User role
        created_at:
          { type: string, format: date-time, description: Creation time }
        updated_at:
          { type: string, format: date-time, description: Update time }
      required:
        [
          uid,
          first_name,
          last_name,
          username,
          email,
          role,
          created_at,
          updated_at,
        ]
...
    APIKey:
      x-entity: true
      type: object
      properties:
        uid: { type: string, description: API Key UID }
        name: { type: string, description: API Key name }
        description:
          { type: string, nullable: true, description: API Key description }
        key_hashed: { type: string, description: Hashed key }
        user:
          $ref: "#/components/schemas/User"
        scopes:
          type: array
          items:
            type: string
          description: List of scopes
        last_used_at:
          {
            type: string,
            format: date-time,
            nullable: true,
            description: Last used time,
          }
        revoked: { type: boolean, description: Is revoked }
        revoked_at:
          { type: string, format: date-time, description: Revocation time }
        expires_at:
          {
            type: string,
            format: date-time,
            nullable: true,
            description: Expiry time,
          }
        created_at:
          { type: string, format: date-time, description: Creation time }
        updated_at:
          { type: string, format: date-time, description: Update time }
      required: [uid, revoked, key_hashed, created_at, updated_at, scopes]
    APIKeyCreate:
      type: object
      properties:
        name:
          type: string
          description: API Key name
        description:
          type: string
          nullable: true
          description: API Key description
        scopes:
          type: array
          items:
            type: string
          description: List of scopes
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: Expiry time
    APIKeyCreateResponse:
      type: object
      properties:
        consumer_key:
          type: string
          description: The consumer key (secret)
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: Expiry time
      required: [consumer_key]
    APIKeyListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/APIKey"
          description: List of API keys
        count:
          type: integer
          description: Total count
      required: [items, count]
...